Refactor (#50)
* rebase changes

* rebase issues fixed

* css lint adjustment

* lint fix

* initial commit

* rename components

* refactor profile page

* responsibility fix

* improved data structure

* repair change name feature

* delete comments

* change git config ignorecase to false

* PR bug fixes
