+++ src/components/Form/styled.js
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 const StyledForm = styled.form`
 	height: 2em;
 	margin: 0 auto;
-	margin: 0.5rem 0 0.5rem 0;
+	margin: 0.5em 0 0.5em 0;
 	input {
 		height: 100%;
 		border: 2px solid var(--dark);
