+++ src/components/NavBar/styled.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const StyledNavBar = styled.nav`
+	display: flex;
+	align-items: center;
+	justify-content: space-evenly;
+	padding: 0.2em 0 0.2em 0;
+	background-color: var(--blue);
+`;
+
+export default StyledNavBar;
