+++ src/components/Profile/styled.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components';
+
+const StyledProfile = styled.section`
+	margin: 0 auto;
+
+	article {
+		display: flex;
+		flex-direction: column;
+		justify-content: center;
+	}
+`;
+
+export default StyledProfile;
