--- src/components/ProfileComponent/styled.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-const StyledProfile = styled.section``;
-
-export default StyledProfile;
