+++ src/components/QuestionCard/styled.js
@@ -4,7 +4,7 @@ import {pxToRem} from '../../utils/unit';
 
 const StyledQuestionComponent = styled.section`
 	position: relative;
-	width: 80vw;
+	width: 90vw;
 	margin: 0.5em 0 0.5em 0;
 	border: 2px solid var(--dark);
 	border-radius: ${pxToRem(8)};
