+++ src/hooks/useStore.js
@@ -1,16 +1,33 @@
 import create from 'zustand';
 import {persist} from 'zustand/middleware';
 
-import questionDatabase from '../utils/questionDatabase';
-import userDatabase from '../utils/userDatabase';
+import dataBase from '../utils/dataBase';
 
 export const useStore = create(
 	persist(
 		set => ({
-			questions: [...questionDatabase],
+			questions: [...dataBase.questions],
+			answers: [...dataBase.answers],
+			users: [...dataBase.users],
+			activeUser: dataBase.users[0],
+
 			setQuestions: newQuestions => {
 				set(() => {
 					return {
+						questions: newQuestions,
+					};
+				});
+			},
+			addNewAnswer: (questionID, newAnswer) => {
+				set(state => {
+					const newQuestions = state.questions.map(question => {
+						if (question.id === questionID) {
+							question.answers = [...question.answers, newAnswer.id];
+						}
+						return question;
+					});
+					return {
+						answers: [newAnswer, ...state.answers],
 						questions: newQuestions,
 					};
 				});
@@ -22,10 +39,18 @@ export const useStore = create(
 					};
 				});
 			},
-			users: userDatabase,
-			activeUser: userDatabase[0],
+			changeUsername: (activeUserID, newUserName) => {
+				set(state => {
+					state.users.map(user => {
+						console.log(user.userID);
+						if (user.userID === activeUserID) {
+							return (user.userName = newUserName);
+						}
+						return null;
+					});
+				});
+			},
 			setActiveUser: newActiveUser => {
-				console.log(newActiveUser);
 				set(() => {
 					return {
 						activeUser: newActiveUser,
