--- src/utils/questionDatabase.js
@@ -1,5 +1,0 @@
-//import {nanoid} from 'nanoid';
-
-const questionDatabase = [];
-
-export default questionDatabase;
