--- src/utils/userDatabase.js
@@ -1,14 +1,0 @@
-import {nanoid} from 'nanoid';
-
-const userDatabase = [
-	{
-		userID: nanoid(),
-		userName: 'Laurenz',
-	},
-	{
-		userID: nanoid(),
-		userName: 'Theo',
-	},
-];
-
-export default userDatabase;
